[AIRFLOW-5502] Move GCP base hook to core
