+++ airflow/contrib/hooks/gdrive_hook.py
@@ -22,7 +22,7 @@ from typing import Any, Optional
 from googleapiclient.discovery import build
 from googleapiclient.http import MediaFileUpload
 
-from airflow.contrib.hooks.gcp_api_base_hook import GoogleCloudBaseHook
+from airflow.gcp.hooks.base import GoogleCloudBaseHook
 
 
 # noinspection PyAbstractClass
