+++ airflow/gcp/hooks/discovery_api.py
@@ -22,7 +22,7 @@ This module allows you to connect to the Google Discovery API Service and query 
 """
 from googleapiclient.discovery import build
 
-from airflow.contrib.hooks.gcp_api_base_hook import GoogleCloudBaseHook
+from airflow.gcp.hooks.base import GoogleCloudBaseHook
 
 
 class GoogleDiscoveryApiHook(GoogleCloudBaseHook):
