+++ airflow/gcp/hooks/gsheets.py
@@ -23,7 +23,7 @@ This module contains a Google Sheets API hook
 
 from typing import Any, Dict, List, Optional
 from googleapiclient.discovery import build
-from airflow.contrib.hooks.gcp_api_base_hook import GoogleCloudBaseHook
+from airflow.gcp.hooks.base import GoogleCloudBaseHook
 from airflow.exceptions import AirflowException
