+++ airflow/gcp/hooks/mlengine.py
@@ -26,7 +26,7 @@ from typing import Dict, Callable, List, Optional
 from googleapiclient.errors import HttpError
 from googleapiclient.discovery import build
 
-from airflow.contrib.hooks.gcp_api_base_hook import GoogleCloudBaseHook
+from airflow.gcp.hooks.base import GoogleCloudBaseHook
 from airflow.utils.log.logging_mixin import LoggingMixin
